+++ supplier_api_tester/supplier_api_tester/tests/reservation.py
@@ -19,7 +19,7 @@ from ..utils.reservation import get_reservation_slot
 
 
 @test_wrapper
-def test_missing_api_key(api_url, api_key, product_id, timeslots: bool, version=2):
+def test_missing_api_key(api_url, api_key, product_id, version=2):
     """Request without API-Key"""
     raw_response, _ = client(f'{api_url}/v{version}/products/{product_id}/reservation', api_key, method=requests.post, json_payload={}, headers={})
 
@@ -42,7 +42,7 @@ def test_missing_api_key(api_url, api_key, product_id, timeslots: bool, version=
 
 
 @test_wrapper
-def test_incorrect_api_key(api_url, api_key, product_id, timeslots: bool, version=2):
+def test_incorrect_api_key(api_url, api_key, product_id, version=2):
     """Request with incorrect API-Key"""
     raw_response, _ = client(f'{api_url}/v{version}/products/{product_id}/reservation', api_key, method=requests.post, json_payload={}, headers={
         'API-Key': 'NON-EXISTING-API-KEY',
@@ -67,11 +67,11 @@ def test_incorrect_api_key(api_url, api_key, product_id, timeslots: bool, versio
 
 
 @test_wrapper
-def test_missing_argument_error(api_url, api_key, product_id, timeslots: bool, version=2):
+def test_missing_argument_error(api_url, api_key, product_id, version=2):
     """Testing missing argument errors"""
     tomorrow = get_tomorrow()
     warnings: List[str] = []
-    slot = get_reservation_slot(api_url, api_key, product_id, timeslots)
+    slot = get_reservation_slot(api_url, api_key, product_id)
 
     # payload is empty
     json_payload = {}
@@ -139,38 +139,30 @@ def test_missing_argument_error(api_url, api_key, product_id, timeslots: bool, v
 
 
 @test_wrapper
-def test_error_for_non_existing_product(api_url, api_key, product_id, timeslots: bool, version=2):
+def test_error_for_non_existing_product(api_url, api_key, product_id, version=2):
     """Testing reservation for non-existing product"""
     url = f'{api_url}/v{version}/products/NON-EXISTING-PRODUCT-ID/reservation'
-    slot = get_reservation_slot(api_url, api_key, product_id, timeslots)
+    slot = get_reservation_slot(api_url, api_key, product_id)
     json_payload = get_payload_for_reservation(api_url, api_key, product_id, slot)
-    if timeslots:
-        json_payload['datetime'] = (
-            f'{slot.date.isoformat()}T{slot.timeslot}'
-            if slot.timeslot != 'no-timeslots'
-            else f'{slot.date.isoformat()}T00:00'
-        )
     raw_response, response = client(url, api_key, method=requests.post, json_payload=json_payload)
     api_error = get_api_error(raw_response, response)
     expected_error = ApiError(
         error_code=1001,
         error='Missing product',
-        message='Product with ID NON-EXISTING-PRODUCT-ID doesn\'t exist',
+        message='Product with ID NON-EXISTING-PRODUCT-ID does not exist',
     )
     check_api_error(raw_response, api_error, expected_error)
     return TestResult()
 
 
 @test_wrapper
-def test_incorrect_date_format(api_url, api_key, product_id, timeslots: bool, version=2):
+def test_incorrect_date_format(api_url, api_key, product_id, version=2):
     """Testing reservation with incorrect date format"""
     url = f'{api_url}/v{version}/products/{product_id}/reservation'
     bad_date_format = '05/05/2020'
-    slot = get_reservation_slot(api_url, api_key, product_id, timeslots)
+    slot = get_reservation_slot(api_url, api_key, product_id)
     json_payload = get_payload_for_reservation(api_url, api_key, product_id, slot)
     json_payload['datetime'] = bad_date_format
-    if timeslots:
-        json_payload['datetime'] = f'{bad_date_format}T{slot.timeslot}'
     raw_response, response = client(url, api_key, method=requests.post, json_payload=json_payload)
     api_error = get_api_error(raw_response, response)
     expected_error = ApiError(
@@ -183,19 +175,13 @@ def test_incorrect_date_format(api_url, api_key, product_id, timeslots: bool, ve
 
 
 @test_wrapper
-def test_past_date(api_url, api_key, product_id, timeslots: bool, version=2):
+def test_past_date(api_url, api_key, product_id, version=2):
     """Testing reservation with past date"""
     url = f'{api_url}/v{version}/products/{product_id}/reservation'
-    slot = get_reservation_slot(api_url, api_key, product_id, timeslots)
+    slot = get_reservation_slot(api_url, api_key, product_id)
     json_payload = get_payload_for_reservation(api_url, api_key, product_id, slot)
     yesterday = datetime.utcnow().date() - timedelta(days=1)
     json_payload['datetime'] = f'{yesterday.isoformat()}T00:00'
-    if timeslots:
-        json_payload['datetime'] = (
-            f'{yesterday.isoformat()}T{slot.timeslot}'
-            if slot.timeslot != 'no-timeslots'
-            else f'{yesterday.isoformat()}T00:00'
-        )
     raw_response, response = client(url, api_key, method=requests.post, json_payload=json_payload)
     api_error = get_api_error(raw_response, response)
     expected_error = ApiError(
@@ -208,17 +194,11 @@ def test_past_date(api_url, api_key, product_id, timeslots: bool, version=2):
 
 
 @test_wrapper
-def test_not_allowed_method(api_url, api_key, product_id, timeslots: bool, version=2):
+def test_not_allowed_method(api_url, api_key, product_id, version=2):
     """Testing methods that are not allowed"""
     url = f'{api_url}/v{version}/products/{product_id}/reservation'
-    slot = get_reservation_slot(api_url, api_key, product_id, timeslots)
+    slot = get_reservation_slot(api_url, api_key, product_id)
     json_payload = get_payload_for_reservation(api_url, api_key, product_id, slot)
-    if timeslots:
-        json_payload['datetime'] = (
-            f'{slot.date.isoformat()}T{slot.timeslot}'
-            if slot.timeslot != 'no-timeslots'
-            else f'{slot.date.isoformat()}T00:00'
-        )
     for method in (requests.get, requests.put, requests.patch, requests.delete):
         raw_response, _ = client(url, api_key, method=method, json_payload=json_payload)
         status_code = getattr(raw_response, 'status_code', 200)
@@ -231,10 +211,10 @@ def test_not_allowed_method(api_url, api_key, product_id, timeslots: bool, versi
 
 
 @test_wrapper
-def test_reservation(api_url, api_key, product_id, timeslots: bool, version=2):
+def test_reservation(api_url, api_key, product_id, version=2):
     """Reserving tickets for at least 1 variant"""
     url = f'{api_url}/v{version}/products/{product_id}/reservation'
-    slot = get_reservation_slot(api_url, api_key, product_id, timeslots)
+    slot = get_reservation_slot(api_url, api_key, product_id)
     json_payload = get_payload_for_reservation(api_url, api_key, product_id, slot)
     raw_response, response = client(url, api_key, method=requests.post, json_payload=json_payload)
     reservation = get_reservation(raw_response, response)