+++ supplier_api_tester/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add support for API v2.0.0
 - drop support for API v1.x
+- drop support for the CLI argument `--timeslots` (`-t`) in v2.0 
 
 ## 1.6.5